Логин работает, добавил ломбок(авто генерация сеттеров, геттеров, пустых конструкторов), можно использовать val для константных списков(которые просто будем передавать в модель без изменения)

+++ src/main/java/kostuchenkov/rgr/data/repository/OrderRepository.java
@@ -3,6 +3,6 @@ package kostuchenkov.rgr.data.repository;
 import kostuchenkov.rgr.data.domain.Order;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Long> {
+public interface OrderRepository extends CrudRepository<Order, Integer> {
 
 }
