+++ src/main/java/kostuchenkov/rgr/data/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package kostuchenkov.rgr.data.repository;
 import kostuchenkov.rgr.data.domain.Review;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReviewRepository extends CrudRepository<Review, Long> {
+public interface ReviewRepository extends CrudRepository<Review, Integer> {
 }
